Fix bug caused by polygons going out of 16 bit range. Corrected indices in portal quad.

+++ CDC/Models/GexModel.cs
@@ -430,7 +430,7 @@ namespace CDC
 		{
 			// Make the vertices unique
 			_geometry.Vertices = new Vertex[_indexCount];
-			for (UInt32 p = 0; p < _polygonCount; p++)
+			for (uint p = 0; p < _polygonCount; p++)
 			{
 				_geometry.Vertices[(3 * p) + 0] = _polygons[p].v1;
 				_geometry.Vertices[(3 * p) + 1] = _polygons[p].v2;
@@ -455,9 +455,9 @@ namespace CDC
 		{
 			MaterialList materialList = null;
 
-			for (UInt16 p = 0; p < _polygonCount; p++)
+			for (uint p = 0; p < _polygonCount; p++)
 			{
-				HandleDebugRendering(p, options);
+				HandleDebugRendering((int)p, options);
 
 				if (materialList == null)
 				{
