+++ CDC/Models/SR1Model.cs
@@ -528,7 +528,7 @@ namespace CDC
 		{
 			// Make the vertices unique
 			_geometry.Vertices = new Vertex[_indexCount];
-			for (UInt32 p = 0; p < _polygonCount; p++)
+			for (uint p = 0; p < _polygonCount; p++)
 			{
 				_geometry.Vertices[(3 * p) + 0] = _polygons[p].v1;
 				_geometry.Vertices[(3 * p) + 1] = _polygons[p].v2;
@@ -618,11 +618,11 @@ namespace CDC
 		{
 			MaterialList materialList = null;
 
-			for (UInt16 p = 0; p < _polygonCount; p++)
+			for (uint p = 0; p < _polygonCount; p++)
 			{
-				ProcessPolygon(p, options);
-				ReadMaterial(reader, p, options);
-				HandleDebugRendering(p, options);
+				ProcessPolygon((int)p, options);
+				ReadMaterial(reader, (int)p, options);
+				HandleDebugRendering((int)p, options);
 
 				if (materialList == null)
 				{
